requirements.txt
asarPy==1.0.1
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
pillow==12.1.1
sqlparse==0.5.5
